+++ gameBoard/src/main/java/com/gameboard/view/controller/SgpostController.java
@@ -1,19 +1,45 @@
 package com.gameboard.view.controller;
 
+import java.io.File;
+import java.io.IOException;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.multipart.MultipartHttpServletRequest;
 
+import com.gameboard.biz.post.SgNoticeService;
 import com.gameboard.biz.post.Sgpost;
+import com.gameboard.biz.post.SgpostComment;
+import com.gameboard.biz.post.SgpostCommentService;
+import com.gameboard.biz.post.SgpostImage;
+import com.gameboard.biz.post.SgpostImageService;
 import com.gameboard.biz.post.SgpostService;
 
 @Controller
 public class SgpostController {
 	@Autowired
 	private SgpostService sg;
+	@Autowired
+	private SgNoticeService noticeService;
+	@Autowired
+	private SgpostImageService ImageService;
+	
+	@Autowired
+	private SgpostCommentService sgCommentService;
 
 	@RequestMapping(value = "getSgID.do")
 	public String getSgID(Model model) {
@@ -21,17 +47,75 @@ public class SgpostController {
 		model.addAttribute("sgDate", sg.getSgDate());
 		return "insertSgpost.jsp";
 	}
-
+	
+		private String maskIpAddress(String ipAddress) {
+		    if (ipAddress.contains(".")) {
+		        String[] parts = ipAddress.split("\\.");
+		        if (parts.length == 4) {
+		            return parts[0] + "." + parts[1] + ".***." + parts[3];
+		        }
+		    }
+		    else if (ipAddress.contains(":")) {
+		        if ("0:0:0:0:0:0:0:1".equals(ipAddress)) {
+		            return "local:01";
+		        } else {
+		            String[] parts = ipAddress.split(":");
+		            return parts[0] + ":" + parts[1] + ":" + parts[2] + ":****:****:" + parts[5] + ":" + parts[6] + ":" + parts[7];
+		        }
+		    }
+		    return ipAddress; 
+		}
+		
+	private String saveFile(MultipartFile file) throws IOException {
+			String fileName = file.getOriginalFilename();
+			String filePath = "resources/images/" + fileName;
+			File destinationFile = new File(filePath);
+			file.transferTo(destinationFile);
+			return fileName;
+	}
+	
 	@RequestMapping(value = "insertSgpost.do")
-	public String insertSgpost(Sgpost vo) {
+	public String insertSgpost(Sgpost vo, HttpServletRequest request, HttpSession session,
+			MultipartHttpServletRequest srequest) {
+		String loggedInMemberId = (String) session.getAttribute("loggedInMemberId");
+
+		if(loggedInMemberId == null) {
+			String ipAddress = request.getRemoteAddr();
+			loggedInMemberId = maskIpAddress(ipAddress);
+		}
+		vo.setUserID(loggedInMemberId);
 		sg.insertSgpost(vo);
+		
+		List<MultipartFile> files = srequest.getFiles("images");
+		for (MultipartFile file : files) {
+			if (!file.isEmpty()) {
+				try {
+					String imageUrl = saveFile(file);
+					SgpostImage image = new SgpostImage();
+					image.setSgID(vo.getSgID());
+					image.setSgImageUrl(imageUrl);
+					ImageService.insertPostImage(image);
+				} catch (IOException e) {
+					e.printStackTrace();
+				}
+			}
+		}
 		return "redirect:suggest.do";
 	}
 
 	@RequestMapping(value = "suggest.do")
 	public String getSgpost(Sgpost vo, Model model) {
+		model.addAttribute("NoticeList", noticeService.getNotices("SG_BOARD_POST"));
 		List<Sgpost> SgList = sg.getSgpostList(vo);
+		
+		Map<Integer, Integer> SGcommentCounts = new HashMap<>();
+		for(Sgpost post : SgList) {
+			int sgCommentCount = sgCommentService.countSgCommentsByPostId(post.getSgID());
+			SGcommentCounts.put(post.getSgID(), sgCommentCount);
+		}
 		model.addAttribute("SgList", SgList);
+		model.addAttribute("SGcommentCounts", SGcommentCounts);
+		
 		return "suggest.jsp";
 	}
 
@@ -42,20 +126,16 @@ public class SgpostController {
 		return "searchSg.jsp";
 	}
 
-	// 수정 + 삭제
 	@RequestMapping(value = "getSgpost.do")
 	public String getSgpostById(int sgID, Model model) {
-		// 조회수 업데이트
 		sg.updateSgpostViews(sgID);
 
 		Sgpost post = sg.getSgpostById(sgID);
 		model.addAttribute("post",post);
 
-		// 이전 게시물과 다음 게시물을 가져오기 위해 ID를 기준으로 조회한다.
-		Sgpost prevPost = sg.getPrevSgpost(sgID); // 이전 게시물 조회
-		Sgpost nextPost = sg.getNextSgpost(sgID); // 다음 게시물 조회
+		Sgpost prevPost = sg.getPrevSgpost(sgID);
+		Sgpost nextPost = sg.getNextSgpost(sgID);
 
-		// 이전 게시물과 다음 게시물이 존재할 경우 모델에 추가한다.
 		if (prevPost != null) {
 			model.addAttribute("prevPost", prevPost);
 		}
@@ -63,30 +143,81 @@ public class SgpostController {
 			model.addAttribute("nextPost", nextPost);
 		}
 
-		// 최신 목록을 가져와서 모델에 추가 (조회수가 업데이트된 상태)
 		List<Sgpost> SgList = sg.getSgpostList(null);
 		model.addAttribute("SgList", SgList);
+		
+		List<SgpostComment> sgCommentList = sgCommentService.getSgCommentsByPostId(sgID);
+		model.addAttribute("sgCommentList", sgCommentList);
+		int SGcommentCounts = sgCommentService.countSgCommentsByPostId(sgID);
+		model.addAttribute("SGcommentCounts", SGcommentCounts);
+		model.addAttribute("sgID", sgID);
 
-		return"getSgpost.jsp"; // 상세 정보를 보여줄 뷰 이름
+		return"getSgpost.jsp";
 	}
 
 
-	@RequestMapping(value = "deleteSgpost.do")
-	public String deleteSgpost(int sgID) {
-		sg.deleteSgpost(sgID);
-		return "redirect:suggest.do";
+	@RequestMapping(value = "deleteSgpost.do", produces = "text/plain; charset=UTF-8")
+	@ResponseBody
+	public String deleteSgpost(int sgID, HttpServletRequest request, HttpSession session) {
+		String loggedInMemberId = (String) session.getAttribute("loggedInMemberId");
+	    if (loggedInMemberId == null) {
+	        String ipAddress = request.getRemoteAddr();
+	        loggedInMemberId = maskIpAddress(ipAddress);
+	    }
+	    Sgpost post = sg.getSgpostById(sgID);
+		
+	    if(post != null && post.getUserID().equals(loggedInMemberId)) {
+	    	sgCommentService.deleteSgAllComment(sgID);
+	    	sg.deleteSgpost(sgID);
+	    	return "deleteSuccess";
+	    }else {
+	    	return "deleteFailed";
+	    }
+	}
+	
+	@RequestMapping(value = "checkEditPermissionSG.do", method = RequestMethod.GET)
+	@ResponseBody
+	public ResponseEntity<String> checkEditPermission(@RequestParam("sgID") int sgID, HttpSession session, HttpServletRequest request) {
+		String loggedInMemberId = (String) session.getAttribute("loggedInMemberId");
+	    if (loggedInMemberId == null) {
+	        String ipAddress = request.getRemoteAddr();
+	        loggedInMemberId = maskIpAddress(ipAddress);
+	    }
+	    
+	    Sgpost post = sg.getSgpostById(sgID);
+	    
+	    if(post != null && post.getUserID().equals(loggedInMemberId)) {
+	    	return ResponseEntity.ok("updateSuccess|updateSgpostForm.do?sgID=" + sgID);
+        } else {
+            return ResponseEntity.ok("updateFailed");
+        }
 	}
 
 	@RequestMapping(value = "updateSgpostForm.do")
-	public String updateSgpostForm(int sgID, Model model) {
-		Sgpost post = sg.getSgpostById(sgID); // 게시물 정보를 가져옴
-		model.addAttribute("post", post); // 수정 폼에서 사용할 게시물 정보를 모델에 추가
-		return "updateSgpostForm.jsp"; // 수정 폼 JSP 페이지로 이동
+	public String updateSgpostForm(@RequestParam("sgID") int sgID, Model model) {
+		List<SgpostImage> images = ImageService.getImagesBySgID(sgID);
+		Sgpost post = sg.getSgpostById(sgID); 
+		model.addAttribute("post", post); 
+		model.addAttribute("images", images);
+		return "updateSgpostForm.jsp"; 
 	}
 
-	@RequestMapping(value = "updateSgpost.do")
-	public String updateSgpost(Sgpost vo) {
-		sg.updateSgpost(vo); // 게시물 정보를 업데이트
+	@RequestMapping(value = "updateSgpost.do", method = RequestMethod.POST)
+	public String updateSgpost(@RequestParam("sgID") int sgID, @ModelAttribute Sgpost vo,
+			HttpSession session, HttpServletRequest request) {
+		String loggedInMemberId = (String) session.getAttribute("loggedInMemberId");
+
+        if (loggedInMemberId == null) {
+            String ipAddress = request.getRemoteAddr();
+            loggedInMemberId = maskIpAddress(ipAddress);
+        }
+        Sgpost existingPost = sg.getSgpostById(sgID);
+        
+        if(existingPost != null && existingPost.getUserID().equals(loggedInMemberId)) {
+        	existingPost.setSgTitle(vo.getSgTitle());
+        	existingPost.setSgContent(vo.getSgContent());
+        	sg.updateSgpost(existingPost);
+        }
 		return "redirect:getSgpost.do?sgID=" + vo.getSgID();
 	}
 }
